requirements.txt
# Web Framework & API
fastapi
streamlit==1.34.0
uvicorn

# Database
motor
pgvector==0.2.5
psycopg[binary]==3.1.18
sqlalchemy==2.0.30
pymongo
neo4j

# Authentication & Security
pydantic==2.7.1
bcrypt
passlib
python-jose

# AI & Machine Learning
openai==1.28.1
phidata==2.4.20
ollama

# Web Tools & Utilities
requests==2.31.0
beautifulsoup4==4.12.3
duckduckgo-search==5.3.1
streamlit-local-storage==0.0.24
google-search-results

# Data Processing
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1

# Visualization
altair==5.3.0
matplotlib==3.8.4

# Development Tools
nest-asyncio==1.6.0
typing-extensions==4.11.0
pydantic-core==2.18.2
pydantic-settings==2.2.1

# PDF Processing
pypdf==4.2.0

# Finance Tools
yfinance==0.2.38

# Utils
python-dateutil==2.9.0.post0
pytz==2024.1
toml==0.10.2
pyyaml==6.0.1

# for backend
fastapi
motor
pydantic
bcrypt
pyjwt
